docs: center avatar on navigation drawer
