+++ packages/api-generator/src/locale/en/VNavigationDrawer.json
@@ -4,7 +4,7 @@
     "clipped": "A clipped drawer rests under the application toolbar. **Note:** requires the **clipped-left** or **clipped-right** prop on `v-app-bar` to work as intended.",
     "disableResizeWatcher": "Prevents the automatic opening or closing of the drawer when resized, based on whether the device is mobile or desktop.",
     "disableRouteWatcher": "Disables opening of navigation drawer when route changes.",
-    "expandOnHover": "Collapses the drawer to a **mini-variant** until hovering with the mouse.",
+    "expandOnHover": "Collapses the drawer to a **rail-variant** until hovering with the mouse.",
     "floating": "A floating drawer has no visible container (no border-right).",
     "height": "Sets the height of the navigation drawer.",
     "image": "Apply a specific background image to the component.",
