+++ packages/docs/src/examples/v-navigation-drawer/misc-combined.vue
@@ -1,23 +1,20 @@
 <template>
   <v-card>
     <v-layout>
-      <v-navigation-drawer
-        theme="dark"
-        permanent
-        rail
-      >
-        <v-list-item
-          prepend-avatar="https://randomuser.me/api/portraits/women/75.jpg"
-          nav
-        ></v-list-item>
+      <v-navigation-drawer theme="dark" permanent rail>
+        <v-list>
+          <v-list-item
+            prepend-avatar="https://randomuser.me/api/portraits/women/75.jpg"
+          ></v-list-item>
+        </v-list>
 
         <v-divider></v-divider>
 
-        <v-list
-          density="compact"
-          nav
-        >
-          <v-list-item prepend-icon="mdi-view-dashboard" value="dashboard"></v-list-item>
+        <v-list density="compact" nav>
+          <v-list-item
+            prepend-icon="mdi-view-dashboard"
+            value="dashboard"
+          ></v-list-item>
 
           <v-list-item prepend-icon="mdi-forum" value="messages"></v-list-item>
         </v-list>
