+++ packages/docs/src/examples/v-navigation-drawer/prop-rail-variant.vue
@@ -1,0 +1,66 @@
+<template>
+  <v-card>
+    <v-layout>
+      <v-navigation-drawer
+        v-model="drawer"
+        :rail="rail"
+        permanent
+        @click="rail = false"
+      >
+        <v-list>
+          <v-list-item
+            prepend-avatar="https://randomuser.me/api/portraits/men/85.jpg"
+            title="John Leider"
+          >
+            <template v-slot:append>
+              <v-btn
+                icon="mdi-chevron-left"
+                variant="text"
+                @click.stop="rail = !rail"
+              ></v-btn>
+            </template>
+          </v-list-item>
+        </v-list>
+
+        <v-divider></v-divider>
+
+        <v-list density="compact" nav>
+          <v-list-item
+            prepend-icon="mdi-home-city"
+            title="Home"
+            value="home"
+          ></v-list-item>
+          <v-list-item
+            prepend-icon="mdi-account"
+            title="My Account"
+            value="account"
+          ></v-list-item>
+          <v-list-item
+            prepend-icon="mdi-account-group-outline"
+            title="Users"
+            value="users"
+          ></v-list-item>
+        </v-list>
+      </v-navigation-drawer>
+      <v-main style="height: 250px"></v-main>
+    </v-layout>
+  </v-card>
+</template>
+
+<script setup>
+  import { ref } from 'vue'
+
+  const drawer = ref(true)
+  const rail = ref(true)
+</script>
+
+<script>
+  export default {
+    data () {
+      return {
+        drawer: true,
+        rail: true,
+      }
+    },
+  }
+</script>
